FWA-14 - Angular code refactoring - cleanup store
