+++ src/app/pages/admin/documents/components/doc-form/doc-form.component.ts
@@ -17,9 +17,9 @@ import { MatOptionModule } from '@angular/material/core';
 import { MatSelectModule } from '@angular/material/select';
 import { Observable } from 'rxjs';
 import { Store } from '@ngrx/store';
-import { selectProjects } from '../../../../../store/selectors/project.selectors';
-import { selectLoading } from '../../../../../store/selectors/document.selectors';
-import { addDocument } from '../../../../../store/actions/document.actions';
+import { selectProjects } from '../../../../../state/project/project.selectors';
+import { selectLoading } from '../../../../../state/document/document.selectors';
+import { addDocument } from '../../../../../state/document/document.actions';
 
 @Component({
   selector: 'app-doc-form',
