+++ src/app/pages/admin/posts/components/post-form/post-form.component.ts
@@ -14,8 +14,8 @@ import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatInputModule } from '@angular/material/input';
 import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
-import { selectLoading } from '../../../../../store/selectors/post.selectors';
-import { addPost } from '../../../../../store/actions/post.actions';
+import { selectLoading } from '../../../../../state/post/post.selectors';
+import { addPost } from '../../../../../state/post/post.actions';
 
 @Component({
   selector: 'app-post-form',
