+++ src/app/state/user/user.selectors.ts
@@ -1,5 +1,5 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
-import { UserState } from '../user.state';
+import { UserState } from './user.reducer';
 
 export const selectUsersState = createFeatureSelector<UserState>('users');
