--- src/app/store/message.state.ts
@@ -1,15 +1,0 @@
-import { Message } from '../interface/message';
-
-export interface MessageState {
-  messages: Message[];
-  isLoaded: boolean;
-  loading: boolean;
-  error: string | null;
-}
-
-export const initialMessagesState: MessageState = {
-  messages: [],
-  isLoaded: false,
-  loading: false,
-  error: null,
-};
